% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mes_fonction.R
\name{filtre_species}
\alias{filtre_species}
\title{Filtrer les observations selon l'espèce}
\usage{
filtre_species(data, Species)
}
\arguments{
\item{data}{Un data.frame contenant une colonne Species}

\item{Species}{Nom de l'espèce à conserver "setosa", "versicolor" ou "virginica"}
}
\value{
Un data.frame filtré selon l'espèce choisie
}
\description{
Cette fonction retourne les lignes d'un jeu de données
correspondant à une espèce donnée de la variable Species.
}
\examples{
filtre_species(iris, "setosa")
filtre_species(iris, "virginica")
}
